+++ src/formatter.js
@@ -1,9 +1,0 @@
-import format from 'prettier-eslint';
-
-export default function formatText({ text, filePath, extensionConfig }) {
-  return format({
-    text,
-    filePath,
-    prettierLast: extensionConfig?.prettierLast || false,
-  });
-}