+++ src/worker/formatter.js
@@ -1,0 +1,13 @@
+module.exports = function formatText({
+  text, prettierEslintPath, filePath, extensionConfig,
+}) {
+  // eslint-disable-next-line import/no-dynamic-require, global-require
+  const format = require(prettierEslintPath);
+
+  return format({
+    text,
+    filePath,
+    logLevel: 'info',
+    prettierLast: extensionConfig?.prettierLast || false,
+  });
+};